move source out to its own module
